Initial implementation of community announcements. Still needs work.
Right now, everyone can edit and delete a post. Only the respective author should be able to do that. Also note that when you click on the title of a post, it will take you to a details page. This might not be necessary for our purposes (Need to talk about this with the group).