Added access control to the loged in navbar.
Made so that items appear to people with the right position. For example: My freights would only appear to drivers, Add user will only appear to managers, etc. Also fixed the announcement page so that only managers, or the person that made the post can delete posts. Also for announcements, made it so that only the people who made the post can edit it. Lastly, changed the Edit page for Announcements so that it matches the "create" page.